Bulk material price lookups
