+++ src/main/java/com/autonomouslogic/everef/industry/IndustryMath.java
@@ -10,7 +10,7 @@ import com.autonomouslogic.everef.model.api.PriceSource;
 import com.autonomouslogic.everef.model.api.SecurityClass;
 import com.autonomouslogic.everef.refdata.BlueprintActivity;
 import com.autonomouslogic.everef.refdata.InventoryType;
-import com.autonomouslogic.everef.service.FuzzworkMarketService;
+import com.autonomouslogic.everef.service.EsiMarketPriceService;
 import com.autonomouslogic.everef.service.MarketPriceService;
 import com.autonomouslogic.everef.util.MathUtil;
 import java.math.BigDecimal;
@@ -21,15 +21,14 @@ import java.util.Optional;
 import java.util.function.Function;
 import javax.inject.Inject;
 import javax.inject.Singleton;
-import lombok.NonNull;
 
 @Singleton
 public class IndustryMath {
 	@Inject
-	protected MarketPriceService marketPriceService;
+	protected EsiMarketPriceService esiMarketPriceService;
 
 	@Inject
-	protected FuzzworkMarketService fuzzworkMarketService;
+	protected MarketPriceService marketPriceService;
 
 	@Inject
 	protected IndustryStructures industryStructures;
@@ -43,7 +42,7 @@ public class IndustryMath {
 	public BigDecimal eiv(BlueprintActivity activity, int runs) {
 		var eiv = BigDecimal.ZERO;
 		for (var material : activity.getMaterials().values()) {
-			var adjPrice = marketPriceService.getEsiAdjustedPrice(material.getTypeId());
+			var adjPrice = esiMarketPriceService.getEsiAdjustedPrice(material.getTypeId());
 			if (adjPrice.isEmpty()) {
 				throw new RuntimeException("typeId: " + material.getTypeId());
 			}
@@ -75,44 +74,12 @@ public class IndustryMath {
 		for (var material : activity.getMaterials().values()) {
 			long typeId = material.getTypeId();
 			var quantity = quantityMod.apply(material.getQuantity());
-			var costPerUnit = materialCostPerUnit(typeId, priceSource);
-			var cost = materialCost(costPerUnit, quantity);
 			materials.put(
 					String.valueOf(typeId),
-					MaterialCost.builder()
-							.typeId(typeId)
-							.quantity(quantity)
-							.costPerUnit(costPerUnit)
-							.cost(cost)
-							.build());
+					MaterialCost.builder().typeId(typeId).quantity(quantity).build());
 		}
-		return materials;
-	}
-
-	public BigDecimal materialCostPerUnit(long typeId, @NonNull PriceSource priceSource) {
-		if (priceSource.isEsi()) {
-			return MathUtil.round(
-					BigDecimal.valueOf(
-							marketPriceService.getEsiAveragePrice(typeId).orElse(0)),
-					2);
-		} else if (priceSource.isFuzzwork()) {
-			var prices = fuzzworkMarketService.fetchAggregateStationPrices(priceSource.getStationId(), List.of(typeId));
-			return Optional.ofNullable(prices.get(String.valueOf(typeId)))
-					.flatMap(p -> switch (priceSource) {
-						case FUZZWORK_JITA_BUY_AVG, FUZZWORK_JITA_BUY_MAX -> Optional.ofNullable(p.getBuy());
-						case FUZZWORK_JITA_SELL_AVG, FUZZWORK_JITA_SELL_MIN -> Optional.ofNullable(p.getSell());
-						default -> throw new RuntimeException(String.format("Unknown price source: %s", priceSource));
-					})
-					.flatMap(p -> switch (priceSource) {
-						case FUZZWORK_JITA_BUY_AVG, FUZZWORK_JITA_SELL_AVG ->
-							Optional.ofNullable(p.getWeightedAverage());
-						case FUZZWORK_JITA_BUY_MAX -> Optional.ofNullable(p.getMax());
-						case FUZZWORK_JITA_SELL_MIN -> Optional.ofNullable(p.getMin());
-						default -> throw new RuntimeException(String.format("Unknown price source: %s", priceSource));
-					})
-					.orElse(BigDecimal.ZERO);
-		}
-		throw new RuntimeException(String.format("Unknown price source: %s", priceSource));
+		var materialsWithCost = marketPriceService.materialCosts(materials, priceSource);
+		return materialsWithCost;
 	}
 
 	public BigDecimal materialCost(BigDecimal price, long quantity) {
