+++ src/main/java/com/autonomouslogic/everef/service/MarketPriceService.java
@@ -1,107 +1,90 @@
 package com.autonomouslogic.everef.service;
 
-import com.autonomouslogic.everef.openapi.esi.api.MarketApi;
-import com.autonomouslogic.everef.openapi.esi.invoker.ApiException;
-import com.autonomouslogic.everef.openapi.esi.invoker.ApiResponse;
-import com.autonomouslogic.everef.openapi.esi.model.GetMarketsPrices200Ok;
+import com.autonomouslogic.everef.model.api.MaterialCost;
+import com.autonomouslogic.everef.model.api.PriceSource;
+import com.autonomouslogic.everef.util.MathUtil;
+import java.math.BigDecimal;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.OptionalDouble;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 import javax.inject.Inject;
 import javax.inject.Singleton;
-import lombok.Data;
-import lombok.SneakyThrows;
-import lombok.extern.log4j.Log4j2;
+import lombok.NonNull;
+import org.apache.commons.lang3.tuple.Pair;
+import org.jetbrains.annotations.NotNull;
 
 @Singleton
-@Log4j2
 public class MarketPriceService {
 	@Inject
-	protected MarketApi marketApi;
+	protected EsiMarketPriceService esiMarketPriceService;
 
 	@Inject
-	protected ScheduledExecutorService scheduler;
-
-	private String esiMarketPricesEtag;
-	private ScheduledFuture<?> marketPricesFuture;
-	private Map<Long, MarketPrice> prices = new ConcurrentHashMap<>();
+	protected FuzzworkMarketService fuzzworkMarketService;
 
 	@Inject
 	protected MarketPriceService() {}
 
-	public void init() {
-		updateEsiMarketPrices();
-		marketPricesFuture = scheduler.scheduleAtFixedRate(
-				() -> {
-					try {
-						updateEsiMarketPrices();
-					} catch (Exception e) {
-						log.warn("Failed to update market prices, ignoring", e);
-					}
-				},
-				10,
-				10,
-				TimeUnit.MINUTES);
-	}
-
-	@SneakyThrows
-	private void updateEsiMarketPrices() {
-		log.info("Updating market prices");
-		ApiResponse<List<GetMarketsPrices200Ok>> res;
-		try {
-			res = marketApi.getMarketsPricesWithHttpInfo(null, esiMarketPricesEtag);
-		} catch (ApiException e) {
-			if (e.getCode() == 304) {
-				log.debug("No market prices update needed");
-				return;
-			} else {
-				throw e;
-			}
+	public Map<Long, BigDecimal> getMarketPrices(@NonNull List<Long> typeIds, @NonNull PriceSource priceSource) {
+		if (priceSource.isEsi()) {
+			return getEsiPrices(typeIds);
+		} else if (priceSource.isFuzzwork()) {
+			getFuzzworkPrices(typeIds, priceSource);
 		}
-		for (GetMarketsPrices200Ok price : res.getData()) {
-			prices.computeIfAbsent(price.getTypeId().longValue(), ignore -> new MarketPrice())
-					.setEsiAdjustedPrice(
-							Optional.ofNullable(price.getAdjustedPrice()).orElse(0.0))
-					.setEsiAveragePrice(
-							Optional.ofNullable(price.getAveragePrice()).orElse(0.0));
-		}
-		esiMarketPricesEtag = res.getHeaders().get("ETag").getFirst();
-		log.info("Finished updating market prices");
+		throw new RuntimeException(String.format("Unknown price source: %s", priceSource));
 	}
 
-	public OptionalDouble getEsiAdjustedPrice(long typeId) {
-		if (!prices.containsKey(typeId)) {
-			return OptionalDouble.empty();
-		}
-		return OptionalDouble.of(prices.get(typeId).getEsiAdjustedPrice());
+	private Map<Long, BigDecimal> getFuzzworkPrices(List<Long> typeIds, PriceSource priceSource) {
+		var mapped = new LinkedHashMap<Long, BigDecimal>();
+		fuzzworkMarketService
+				.fetchAggregateStationPrices(priceSource.getStationId(), typeIds)
+				.forEach((key, type) -> {
+					var price = Optional.ofNullable(type)
+							.flatMap(p -> switch (priceSource) {
+								case FUZZWORK_JITA_BUY_AVG, FUZZWORK_JITA_BUY_MAX -> Optional.ofNullable(p.getBuy());
+								case FUZZWORK_JITA_SELL_AVG, FUZZWORK_JITA_SELL_MIN -> Optional.ofNullable(p.getSell());
+								default ->
+									throw new RuntimeException(String.format("Unknown price source: %s", priceSource));
+							})
+							.flatMap(p -> switch (priceSource) {
+								case FUZZWORK_JITA_BUY_AVG, FUZZWORK_JITA_SELL_AVG ->
+									Optional.ofNullable(p.getWeightedAverage());
+								case FUZZWORK_JITA_BUY_MAX -> Optional.ofNullable(p.getMax());
+								case FUZZWORK_JITA_SELL_MIN -> Optional.ofNullable(p.getMin());
+								default ->
+									throw new RuntimeException(String.format("Unknown price source: %s", priceSource));
+							})
+							.orElse(BigDecimal.ZERO);
+					mapped.put(Long.valueOf(key), price);
+				});
+		return mapped;
 	}
 
-	public OptionalDouble getEsiAveragePrice(long typeId) {
-		if (!prices.containsKey(typeId)) {
-			return OptionalDouble.empty();
-		}
-		return OptionalDouble.of(prices.get(typeId).getEsiAveragePrice());
+	private @NotNull Map<Long, BigDecimal> getEsiPrices(@NonNull List<Long> typeIds) {
+		return typeIds.stream()
+				.map(typeId -> Pair.of(
+						typeId,
+						MathUtil.round(
+								BigDecimal.valueOf(esiMarketPriceService
+										.getEsiAveragePrice(typeId)
+										.orElse(0)),
+								2)))
+				.collect(Collectors.toMap(Pair::getLeft, Pair::getRight));
 	}
 
-	public boolean isReady() {
-		return !prices.isEmpty();
-	}
-
-	public void stop() {
-		if (marketPricesFuture != null) {
-			marketPricesFuture.cancel(true);
-			marketPricesFuture = null;
-		}
-	}
-
-	@Data
-	private static final class MarketPrice {
-		private double esiAdjustedPrice;
-		private double esiAveragePrice;
+	public Map<String, MaterialCost> materialCosts(Map<String, MaterialCost> materials, PriceSource priceSource) {
+		var typeIds = materials.keySet().stream().map(Long::valueOf).toList();
+		var prices = getMarketPrices(typeIds, priceSource);
+		var newMaterials = new LinkedHashMap<String, MaterialCost>();
+		materials.keySet().forEach(key -> {
+			var material = materials.get(key);
+			var costPerUnit = prices.getOrDefault(Long.valueOf(key), BigDecimal.ZERO);
+			var cost = MathUtil.round(costPerUnit.multiply(BigDecimal.valueOf(material.getQuantity())), 2);
+			newMaterials.put(
+					key,
+					material.toBuilder().costPerUnit(costPerUnit).cost(cost).build());
+		});
+		return newMaterials;
 	}
 }
